+++ flask_unchained/bundles/sqlalchemy/alembic/__init__.py
@@ -1,1 +1,0 @@
-from .materialized_view import MaterializedViewMigration