+++ tests/bundles/sqlalchemy/services/test_session_manager.py
@@ -1,7 +1,7 @@
 from flask_unchained.bundles.sqlalchemy import SessionManager, SQLAlchemyUnchained
 
 
-def setup(db: SQLAlchemyUnchained):
+def _setup(db: SQLAlchemyUnchained):
     session_manager = SessionManager(db)
 
     class Foo(db.Model):
@@ -16,7 +16,7 @@ def setup(db: SQLAlchemyUnchained):
 
 class TestSessionManager:
     def test_save(self, db: SQLAlchemyUnchained):
-        Foo, session_manager = setup(db)
+        Foo, session_manager = _setup(db)
 
         foo = Foo(name='foo')
         session_manager.save(foo)
@@ -24,14 +24,14 @@ class TestSessionManager:
         # check it's added to the session but not committed
         assert foo in db.session
         with db.session.no_autoflush:
-            assert Foo.q.get_by(name='foo') is None
+            assert Foo.q.filter_by(name='foo').one_or_none() is None
 
         # check the commit kwarg works
         session_manager.save(foo, commit=True)
-        assert Foo.q.get_by(name='foo') == foo
+        assert Foo.q.filter_by(name='foo').one() == foo
 
     def test_save_all(self, db: SQLAlchemyUnchained):
-        Foo, session_manager = setup(db)
+        Foo, session_manager = _setup(db)
 
         foo1 = Foo(name='one')
         foo2 = Foo(name='two')
@@ -42,14 +42,14 @@ class TestSessionManager:
         with db.session.no_autoflush:
             for foo in all_:
                 assert foo in db.session
-                assert Foo.q.get_by(name=foo.name) is None
+                assert Foo.q.filter_by(name=foo.name).one_or_none() is None
 
         session_manager.save_all(all_, commit=True)
         for foo in all_:
-            assert Foo.q.get_by(name=foo.name) == foo
+            assert Foo.q.filter_by(name=foo.name).one() == foo
 
     def test_delete(self, db: SQLAlchemyUnchained):
-        Foo, session_manager = setup(db)
+        Foo, session_manager = _setup(db)
 
         foo1 = Foo(name='one')
         foo2 = Foo(name='two')
@@ -58,10 +58,10 @@ class TestSessionManager:
 
         for foo in all_:
             assert foo in db.session
-            assert Foo.q.get_by(name=foo.name) == foo
+            assert Foo.q.filter_by(name=foo.name).one() == foo
 
         session_manager.delete(foo1, commit=True)
         assert foo1 not in db.session
-        assert Foo.q.get_by(name='one') is None
+        assert Foo.q.filter_by(name='one').one_or_none() is None
         assert foo2 in db.session
-        assert Foo.q.get_by(name='two') == foo2
+        assert Foo.q.filter_by(name='two').one() == foo2