+++ tests/bundles/sqlalchemy/test_custom_extension.py
@@ -14,7 +14,7 @@ from tests.bundles.sqlalchemy._bundles.custom_extension.extensions import (
 @pytest.mark.usefixtures('app', 'db')
 class TestCustomExtension:
     def test_it_works(self, app, db):
-        exts = [db, app.extensions['sqlalchemy'].db, unchained.extensions.db]
+        exts = [db, app.extensions['sqlalchemy'], unchained.extensions.db]
         for i, ext in enumerate(exts):
             if is_local_proxy(ext):
                 ext = ext._get_current_object()
@@ -23,7 +23,6 @@ class TestCustomExtension:
 
     def test_it_uses_the_correct_base_model(self, db):
         assert issubclass(db.Model, CustomModel)
-        assert issubclass(db.MaterializedView, CustomModel)
 
         from ._bundles.ext_vendor_one.models import OneBasic, OneRole
         assert issubclass(OneBasic, CustomModel)
@@ -35,9 +34,4 @@ class TestCustomExtension:
     def test_base_model_meta_options_are_correct(self, db):
         assert db.Model.Meta.extend_existing is True
         assert db.Model.Meta.pk == 'pk'
-        assert db.Model.Meta._testing_ == 'overriding the default'
-
-        assert db.MaterializedView.Meta.pk is None
-        assert db.MaterializedView.Meta.created_at is None
-        assert db.MaterializedView.Meta.updated_at is None
         assert db.Model.Meta._testing_ == 'overriding the default'